Adjust client package unit tests.
The listing test only needs to make sure there's at least one key
in the datastore (the "foo" key that the test makes sure is present).
With more dev work happening, keys are being added and it's superfluous
to have to reset the datastore to a known state before testing.